+++ lib/capybara_accessible_selectors/selectors/region.rb
@@ -3,17 +3,23 @@
 Capybara.add_selector :region, locator_type: [String, Symbol] do
   xpath do |*|
     XPath.descendant[[
-      XPath.local_name == "section",
+      XPath.self(:section) & (XPath.attr(:"aria-label") | XPath.attr(:"aria-labelledby")),
       XPath.attr(:role) == "region"
     ].reduce(:|)]
   end
 
-  locator_filter skip_if: nil do |node, locator, exact:, **|
-    method = exact ? :eql? : :include?
-    if node[:"aria-labelledby"]
-      CapybaraAccessibleSelectors::Helpers.element_labelledby(node).public_send(method, locator)
-    elsif node[:"aria-label"]
-      node[:"aria-label"].public_send(method, locator.to_s)
+  locator_filter do |node, locator, exact:, **|
+    next true if !locator && node.tag_name != "section"
+
+    name = if node[:"aria-labelledby"]
+             CapybaraAccessibleSelectors::Helpers.element_labelledby(node)
+           else
+             node[:"aria-label"]
+           end
+    if locator
+      name&.public_send(exact ? :eql? : :include?, locator)
+    else
+      name && !name.strip.empty?
     end
   end
 
@@ -26,8 +32,8 @@ module CapybaraAccessibleSelectors
     #
     # @param [String] Name Region label
     # @param [Hash] options Finder options
-    def within_region(*arguments, **options, &block)
-      within(:region, *arguments, **options, &block)
+    def within_region(...)
+      within(:region, ...)
     end
   end
 end