+++ spec/spec_helper.rb
@@ -2,6 +2,9 @@
 
 ENV["APP_ENV"] = "test"
 
+Warning[:deprecated] = true
+Warning[:performance] = true if RUBY_VERSION >= "3.3.0"
+
 require "debug"
 require "capybara/rspec"
 require "selenium-webdriver"
@@ -31,15 +34,7 @@ class String
 end
 
 driver = ENV["DRIVER"]&.to_sym || :selenium_chrome_headless
-
 Capybara.register_driver(:safari) { |app| Capybara::Selenium::Driver.new(app, browser: :safari) }
-Capybara.register_driver(:firefox_developer_edition) do |app|
-  options = Selenium::WebDriver::Firefox::Options.new(
-    binary: "/Applications/Firefox Developer Edition.app/Contents/MacOS/firefox-bin"
-  )
-  options.headless!
-  Capybara::Selenium::Driver.new(app, browser: :firefox, capabilities: options)
-end
 Capybara.default_driver = driver
 Capybara.app = CapybaraAccessibleSelectors::TestApplication
 Capybara.server = :puma, { Silent: true }
@@ -60,4 +55,26 @@ RSpec.configure do |config|
       visit("/pages/new?body=#{CGI.escape(html.strip)}")
     end
   end)
+end
+
+module Capybara
+  module Node
+    module WarnOnTimeouts
+      def synchronize(seconds = nil, *, **)
+        start_time = Time.now
+        super
+      rescue Capybara::ElementNotFound => e
+        seconds ||= Capybara.default_max_wait_time
+        unless seconds.zero? || Time.now - start_time <= seconds
+          stack = caller.select { _1.start_with?(Dir.pwd) }.map { _1.slice(Dir.pwd.length..) }.drop(1)
+          warn("selector has timed out #{stack}")
+        end
+        raise e
+      end
+    end
+
+    class Base
+      prepend WarnOnTimeouts
+    end
+  end
 end